Add an app for e2e test on the runtime builder (#85)
The app is built and deployed through the new runtime builder build steps. The tests check if the app can run and interact with Stackdriver APIs. The tests also contains two custom tests, one uses the net package which uses CGO, another prints the Go version we use to build the app.